Split Map to its own file, store MSBs and msbIds in bank, locator fixes
