+++ src/StudioCore/Tests/GlyphCatcher.cs
@@ -34,7 +34,7 @@ public static class GlyphCatcher
     /// </summary>
     public static unsafe void CheckMSB(AssetLocator locator)
     {
-        var maps = locator.GetFullMapList();
+        var maps = Locator.ActiveProject.MSBBank.GetFullMapList();
         HashSet<char> msbChars = new();
         foreach (var mapName in maps)
         {
